Trying out new ibcdfo python package structure.
This structure allows for subpackages, which is as desired.  In addition,
pounders functions are imported automatically into the ibcdfo.pounders
namespace.  I also converted some python comments to python docstrings so that
the package's documentation is full.  Left a few TODOs for Jeff before this can
be merged into main.

I have not tested these changes at all.  I have just inspected the contents of
the sdist tarball to confirm that it contains the pounders *.py files as well as
pounders test python files.  It also includes the LICENSE and README.md files
from the root of the repository.  Also, I confirmed that the LICENSE file is
actually installed along with the source when the sdist tarball is used to
install with pip.  Confirmed that I can use pydoc to see all package
documentation as expected.