Use relative imports within package.
In at least one test run locally, this got rid of an error that claimed that
pounders was calling a module instead of what should be a function.  Replaced
use of pkg_resources with importlib since pip complained that the former is
deprecated.  Finally, "import ibcdfo" will now import pounders automatically so
that it can be used immediately.