(Issue #145) See if we need special git settings.
I am getting a Commit creating failed: ["Service not found: none"] failure in
CodeCov.  Some issues online make it sound like this is due to git not having
the correct safe directory set.  The checkout action is setting safe
directories to both of the repos.  However, these are not being set for the
default folder due to my custom installation of two repos.  Setting the default
workspace as safe to see if this helps.  This is where the reports are and
where code cov is running from (AFAIK).