Update README.md
WIP - Need to expand to more clearly allow for F_aux to be returned from Ffun. For now this is not described (it would in Outputs Model, returned outputs, Ffun description, etc.).

The present change is simply to align the language of F to make this consistent with potential change to Ffun.