(Issue #173) Don't include historic code in distributions.
Also added READMEs so that people can understand why the old code is still in
the repo.  The test_phi2eval test was in two places.  Jeff wants it to be
clearly indicated as a unit test, so we pulled it out of the test case of mostly
end-to-end tests.

I generated a source distribution and a wheel and confirmed that the
old_unit_tests and unused_code folders are no longer in them.