(Issue #217) See if we can get actions to use public POptUS repo.
Since that repo is a modern version, it has a slightly different interface
including new class names.  Had to tweak pounders.py to get it using the logging
facilities.

Note that POUNDers unittests are presently failing due to the need to update
concurrent POUNDers in accord with other changes made on this branch.

tox tasks were passing locally aside from the above.