(Issue #227) Eagerly install old jax/jaxlib packages.
During PR review, local testing revealed that the oldest tox task was not
running the IBCDFO test with the desired oldest numpy/scipy packages.  This
appears to happen because pip was installing a new-ish version of jax that
required newer versions of numpy/scipy, which is inline with pyproject.toml.

I want to see if this workaround results in GH action jobs using appropriate
numpy/scipy versions.