(Issue #226) Correctly distinguish between the "smooth" and "nonsmooth" hfuns.
This is only necessary for MSP and what "smooth" means will be clarified in the
future in the main MSP docs.  The inclusion of the least squares h function is
for testing and there is no need to include it in the public interface.
Therefore, the `general_smooth_h_functions` folder is not mentioned in the
MATLAB docs.