(Issue #226) Smooth hfuns should not be in MATLAB public interface for MSP.
If users have a "smooth" hfun, then they should likely be using a different
method.  Therefore, do not give users the false impression that they could use
the smooth hfun that exists for testing only.

Note that this same change does not need to be made in the Python
implementation yet since that test, and therefore the least-squares hfun, don't
exist in the Python implementation.